+++ packages/approval_tests/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+- Breaking release:
+   - Added new reporter: `GitReporter`. It allows you to use `git` to view the differences between the received and approved files.
+   - Added support to approve files using CLI. Now you can approve files using the command line: `dart run approved:review`
+   - Added support to use ApprovalTests during widget tests.  
+Thanks to [Richard Coutts](https://github.com/buttonsrtoys)
+
 ## 1.0.0
 
 - Initial major release.